requirements.txt
fastapi
youtube_dl==2021.2.10
uvicorn
loguru==0.5.3
aiofiles==0.6.0
python-telegram-bot==13.2
py_de_familia==0.1.3